bug(signup-validation): signup validation rule failing
- Fix signup validation rule failing

[Finishes: #156891044]
