+++ server/v2/models/index.js
@@ -9,7 +9,6 @@ dotenv.config();
 const env = process.env.NODE_ENV || 'development';
 const config = require(`${__dirname}/../config/config.js`)[env];
 const db = {};
-
 let sequelize;
 
 if (config.use_env_variable) {
